pyproject.toml
[project]
name = "green-ctx"
version = "0.1.0"
requires-python = ">=3.11"
description = "GPU Context Multiplexer"
authors = [
    {name = "Green Context Team"}
]
dependencies = [
    "typer>=0.9.0",
    "pyzmq>=25.1.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "cuda-python",
    "nvmath-python",
    "torch",
    "grpcio==1.59.3",
    "grpcio-tools==1.59.3",
    "protobuf==4.25.1",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.scripts]
gtx = "green_ctx.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.12.1",
    "pytest>=7.4.4",
]

[tool.hatch.metadata]
allow-direct-references = true
